Dockerize application (#31)
* dockerized application

* Added basic docker docs

+++ shoppinglist-be/src/app/database.ts
@@ -1,11 +1,20 @@
 import Database from "better-sqlite3";
 import { Item } from "./types";
+import { mkdirSync } from "fs";
 
 export class DatabaseWrapper {
-  private DBSOURCE = "db.sqlite";
-  private database = new Database(this.DBSOURCE, { verbose: console.log });
+  private dbPath = "data";
+  private database?: Database.Database;
 
   constructor() {
+    // create data dir if nor already existent
+    console.log("creating data dir");
+    mkdirSync(this.dbPath, { recursive: true });
+
+    this.database = new Database(this.dbPath + "/db.sqlite", {
+      verbose: console.log,
+    });
+
     const createTableStmt = `CREATE TABLE IF NOT EXISTS shoppinglist (
                     'id' INTEGER PRIMARY KEY AUTOINCREMENT,
                     'name' text UNIQUE,
@@ -17,6 +26,10 @@ export class DatabaseWrapper {
   }
 
   addItem(item: Item) {
+    if (!this.database) {
+      this.handleError("noDb");
+      return;
+    }
     const stmt = this.database
       .prepare(
         `INSERT INTO shoppinglist (name, amount, onShoppinglist)
@@ -27,6 +40,10 @@ export class DatabaseWrapper {
   }
 
   getItems() {
+    if (!this.database) {
+      this.handleError("noDb");
+      return;
+    }
     const items: Item[] = this.database
       .prepare(
         `SELECT *
@@ -37,6 +54,10 @@ export class DatabaseWrapper {
   }
 
   updateItem(id: number, item: Item) {
+    if (!this.database) {
+      this.handleError("noDb");
+      return;
+    }
     this.database
       .prepare(
         `UPDATE shoppinglist 
@@ -47,11 +68,22 @@ export class DatabaseWrapper {
   }
 
   deleteItem(id: number) {
+    if (!this.database) {
+      this.handleError("noDb");
+      return;
+    }
     this.database
       .prepare(
         `DELETE FROM shoppinglist
     WHERE id = ?`
       )
       .run(id);
+  }
+
+  private handleError(type: "noDb") {
+    switch (type) {
+      case "noDb":
+        throw "database not defined";
+    }
   }
 }
